backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
google-generativeai==0.8.0
requests==2.31.0
aiofiles==0.24.0